Move CESK module from swarm-scenario to swarm-engine
