+++ src/Swarm/TUI/View/CellDisplay.hs
@@ -20,7 +20,6 @@ import Data.Tagged (unTagged)
 import Data.Word (Word32)
 import Graphics.Vty qualified as V
 import Linear.Affine ((.-.))
-import Swarm.Game.CESK (TickNumber (..))
 import Swarm.Game.Display (
   Attribute (AEntity),
   Display,
@@ -39,6 +38,7 @@ import Swarm.Game.State
 import Swarm.Game.State.Robot
 import Swarm.Game.State.Substate
 import Swarm.Game.Terrain
+import Swarm.Game.Tick (TickNumber (..))
 import Swarm.Game.Universe
 import Swarm.Game.World qualified as W
 import Swarm.TUI.Editor.Masking
