+++ src/swarm-engine/Swarm/Game/Exception.hs
