+++ src/swarm-scenario/Swarm/Game/Robot.hs
@@ -16,6 +16,7 @@ module Swarm.Game.Robot (
   _robotID,
   _robotLocation,
   _robotContext,
+  RobotMachine,
   _machine,
   emptyActivityCount,
 
@@ -25,9 +26,6 @@ module Swarm.Game.Robot (
   RobotR,
   Robot,
   TRobot,
-
-  -- ** Runtime robot update
-  C.RobotUpdate (..),
 
   -- * Robot context
   WalkabilityContext (..),
@@ -53,7 +51,6 @@ module Swarm.Game.Robot (
   robotID,
   robotParentID,
   robotHeavy,
-  machine,
   systemRobot,
   selfDestruct,
   runningAtomic,
@@ -69,10 +66,6 @@ module Swarm.Game.Robot (
 
   -- ** Query
   robotKnows,
-  isActive,
-  wantsToStep,
-  waitingUntil,
-  getResult,
 
   -- ** Constants
   hearingDistance,
@@ -84,7 +77,6 @@ import Data.Aeson qualified as Ae (FromJSON, ToJSON (..))
 import Data.Hashable (hashWithSalt)
 import Data.Kind qualified
 import Data.Map (Map)
-import Data.Maybe (isNothing)
 import Data.Sequence (Seq)
 import Data.Sequence qualified as Seq
 import Data.Set (Set)
@@ -92,16 +84,15 @@ import Data.Text (Text)
 import Data.Yaml (FromJSON (parseJSON), (.!=), (.:), (.:?))
 import GHC.Generics (Generic)
 import Linear
-import Swarm.Game.CESK qualified as C
 import Swarm.Game.Display (Display, curOrientation, defaultRobotDisplay, invisible)
 import Swarm.Game.Entity hiding (empty)
 import Swarm.Game.Location (Heading, Location, toDirection, toHeading)
+import Swarm.Game.Tick
 import Swarm.Game.Universe
 import Swarm.Language.Capability (Capability)
 import Swarm.Language.Pipeline (ProcessedTerm)
 import Swarm.Language.Syntax (Const, Syntax)
 import Swarm.Language.Text.Markdown (Document)
-import Swarm.Language.Value as V
 import Swarm.Log
 import Swarm.Util.Lens (makeLensesExcluding, makeLensesNoSigs)
 import Swarm.Util.WindowedCounter
@@ -125,7 +116,7 @@ data ActivityCounts = ActivityCounts
   , _tangibleCommandCount :: Int
   , _commandsHistogram :: Map Const Int
   , _lifetimeStepCount :: Int
-  , _activityWindow :: WindowedCounter C.TickNumber
+  , _activityWindow :: WindowedCounter TickNumber
   }
   deriving (Eq, Show, Generic, Ae.FromJSON, Ae.ToJSON)
 
@@ -197,7 +188,7 @@ commandsHistogram :: Lens' ActivityCounts (Map Const Int)
 lifetimeStepCount :: Lens' ActivityCounts Int
 
 -- | Sliding window over a span of ticks indicating ratio of activity
-activityWindow :: Lens' ActivityCounts (WindowedCounter C.TickNumber)
+activityWindow :: Lens' ActivityCounts (WindowedCounter TickNumber)
 
 -- | With a robot template, we may or may not have a location.  With a
 --   concrete robot we must have a location.
@@ -212,7 +203,6 @@ type family RobotID (phase :: RobotPhase) :: Data.Kind.Type where
 
 type family RobotMachine (phase :: RobotPhase) :: Data.Kind.Type
 type instance RobotMachine 'TemplateRobot = Maybe ProcessedTerm
-type instance RobotMachine 'ConcreteRobot = C.CESK
 
 type family RobotContextMember (phase :: RobotPhase) :: Data.Kind.Type
 type instance RobotContextMember 'TemplateRobot = ()
@@ -249,7 +239,7 @@ deriving instance (Eq (RobotLocation phase), Eq (RobotID phase), Eq (RobotMachin
 -- See https://byorgey.wordpress.com/2021/09/17/automatically-updated-cached-views-with-lens/
 -- for the approach used here with lenses.
 
-makeLensesExcluding ['_robotCapabilities, '_equippedDevices, '_robotLog, '_robotContext] ''RobotR
+makeLensesExcluding ['_robotCapabilities, '_equippedDevices, '_robotLog, '_robotContext, '_machine] ''RobotR
 
 -- | A template robot, i.e. a template robot record without a unique ID number,
 --   and possibly without a location.
@@ -405,9 +395,6 @@ robotKnows r e = contains0plus e (r ^. robotInventory) || contains0plus e (r ^. 
 robotCapabilities :: Getter Robot (Set Capability)
 robotCapabilities = to _robotCapabilities
 
--- | The robot's current CEK machine state.
-machine :: Lens' Robot C.CESK
-
 -- | Is this robot a "system robot"?  System robots are generated by
 --   the system (as opposed to created by the user) and are not
 --   subject to the usual capability restrictions.
@@ -523,31 +510,6 @@ instance FromJSONE EntityMap TRobot where
       <*> liftE (v .:? "heavy" .!= False)
       <*> liftE (v .:? "unwalkable" ..!= mempty)
       <*> pure 0
-
--- | Is the robot actively in the middle of a computation?
-isActive :: Robot -> Bool
-{-# INLINE isActive #-}
-isActive = isNothing . getResult
-
--- | "Active" robots include robots that are waiting; 'wantsToStep' is
---   true if the robot actually wants to take another step right now
---   (this is a /subset/ of active robots).
-wantsToStep :: C.TickNumber -> Robot -> Bool
-wantsToStep now robot
-  | not (isActive robot) = False
-  | otherwise = maybe True (now >=) (waitingUntil robot)
-
--- | The time until which the robot is waiting, if any.
-waitingUntil :: Robot -> Maybe C.TickNumber
-waitingUntil robot =
-  case _machine robot of
-    C.Waiting time _ -> Just time
-    _ -> Nothing
-
--- | Get the result of the robot's computation if it is finished.
-getResult :: Robot -> Maybe (Value, C.Store)
-{-# INLINE getResult #-}
-getResult = C.finalValue . view machine
 
 hearingDistance :: (Num i) => i
 hearingDistance = 32
