+++ test/unit/TestScoring.hs
@@ -6,11 +6,11 @@ module TestScoring where
 import Data.Text.IO qualified as TIO
 import Data.Time.Calendar.OrdinalDate
 import Data.Time.LocalTime
-import Swarm.Game.CESK (TickNumber (..))
 import Swarm.Game.Scenario.Scoring.Best
 import Swarm.Game.Scenario.Scoring.CodeSize
 import Swarm.Game.Scenario.Scoring.ConcreteMetrics
 import Swarm.Game.Scenario.Scoring.GenericMetrics
+import Swarm.Game.Tick (TickNumber (..))
 import Swarm.Language.Module
 import Swarm.Language.Pipeline
 import Swarm.Language.Syntax
